+++ Locking Design Doc.md
@@ -67,42 +67,26 @@ While a stream is locked, it is indistinguishable from a stream that has been dr
 
 For readers, you should be able to tell if they're still active (i.e. have not been released) via `reader.isActive`.
 
+Once a reader is released, it behaves like a closed stream (unless the encapsulated stream has already errored, in which case it behaves like the errored stream).
+
+Note that with this setup, all the same invariants apply to readable streams as they do to readers. For example, when `ready` is fulfilled, the reader's `state` property will no longer return `"waiting"`, and `read()` will return a chunk, just like with a stream.
+
 ### Level 2: subclassers of `ReadableStream`
 
-Subclasses of `ReadableStream` should get locking support "for free." The same mechanisms for acquiring and using a lock should work flawlessly. More interestingly, if they wanted to support modifying the behavior of e.g. `read()` (or `state` or `ready` or `closed`), they should only have to override it in one location.
+Subclasses of `ReadableStream` should get locking support "for free," within reason. The same mechanisms for acquiring and using a lock should work flawlessly.
 
-Which location is more friendly? Probably in `ReadableStream`, so that `ExclusiveStreamReader` still works for `ReadableStream` subclasses. Less work.
+However, if the subclasser starts overriding `read()`, `state`, or `ready`, they will be in trouble. These are delicate operations that reflect the state of the internal queue. The point of the `ExclusiveStreamReader` is to bypass the developer's ability to directly inspect the internal queue.
 
-This means `ExclusiveStreamReader` should delegate to `ReadableStream`, and not the other way around.
+As such, we design `read()`, `state`, and `ready` for readers to bypass the public API of the readable stream and go directly to its internal queue. That means any subclass customizations to `read()` et al. will be bypassed when using `ExclusiveStreamReader`.
 
-### Level 3: custom readable stream implementations?
+In most cases we can imagine, this will be exactly what you want. For example, let's say you were trying to do something like Node.js streams, which emit a `"data"` event whenever a chunk is read. One way of doing this (perhaps not the best way) would be to subclass `ReadableStream` and replace the `read` method with something like `this.emit("data", chunk); super.read(chunk);`. But this is exactly the kind of code an exclusive reader should bypass!
 
-It is unclear whether this is necessary, but up until now we have a high level of support for anyone who wants to re-implement the entire `ReadableStream` interface with their own specific code. For example, if you implement `state`, `ready`, `closed`, `read()`, and `cancel()`, you can do `myCustomStream.pipeTo = ReadableStream.prototype.pipeTo` and it will continue to work.
+However, if you really need your exclusive reader to work with customized `read()`, `state`, or `ready`, there's an escape hatch! You can just implement your own reader class, and return it from your overridden `getReader()` method. It can coordinate with your overridden `read()`/`state`/`ready` all it wants.
 
-If we encourage this kind of thing, we should make it easy for custom readable streams to be lockable as well. That basically means `ExclusiveStreamReader` should not require knowledge of `ReadableStream`'s internal slots.
+### Level 3: custom readable stream implementations
 
-We can work around this if necessary by passing `ExclusiveStreamReader` any capabilities it needs to manipulate `ReadableStream`'s internal state; then people reimplementing the readable stream interface can do e.g. `new ExclusiveStreamReader(this, { getLock, setLock })` or similar.
+A custom readable stream implementation is a class that intends to behave like a readable stream, but does not subclass `ReadableStream` per se. A good example is the experimental `ReadableByteStream` we have in this repo.
 
-## Optimizability
+As explained above, readers are coupled fairly tightly to the internal queue of the stream. Since custom readable streams can have an arbitrary internal structure, custom readable streams will need to implement their own readers and `getReader()` method.
 
-The need to support subclassing, via `ExclusiveStreamReader` delegating to the `ReadableStream` implementation, conflicts a bit with the desire for readers to be fast. However, this can be fixed with some cleverness.
-
-The spec semantics for e.g. `reader.read()` are essentially:
-
-- Check that `reader@[[stream]]` is locked to `reader`.
-- Unlock `reader@[[stream]]`.
-- Try `return reader@[[stream]].read()`; finally re-lock `reader@[[stream]]`.
-
-This will ensure that if `reader@[[stream]]` is a subclass of `ReadableStream`, it will polymorphically dispatch to the subclass's `read` method. However, this kind of try/finally pattern is not very optimizable in V8.
-
-Here is an optimization that can be performed instead, with slight tweaks to both `ReadableStream.prototype.read` and `ExclusiveStreamReader.prototype.read`:
-
-- Define `ReadableStream.prototype.read` as:
-    - Check that `this` is not locked.
-    - Return `ReadFromReadableStream(this)`. (That is, extract the main functionality, without the check, into its own function.)
-- Define `ExclusiveStreamReader.prototype.read` like so:
-    - Check that `this@[[stream]]` is locked to `this`.
-    - If `this@[[stream]].read` is equal to the original `ReadableStream.prototype.read`: return `ReadFromReadableStream(this@[[stream]])`.
-    - Otherwise, proceed via the per-spec semantics above.
-
-This essentially ensures that all undisturbed readable streams, or readable stream subclasses that do not override `read`, go down the "fast path" by ignoring all the try/finally and lock/unlock business. It is unobservable, since we have checked that `read` has not been modified in any way.
+Note that if they implement a `getReader()` that returns something conforming to the exclusive stream reader API, along with `state`, `ready`, `closed`, `read()`, and `cancel()`, then `ReadableStream.prototype.pipeTo` (and `pipeThrough`) will still work when applied generically to the custom readable stream.
